+++ ios/Classes/MyMapView.swift
@@ -24,14 +24,21 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
     var isFollowUserLocation: Bool = false
     var canGetLastUserLocation = false
     var canTrackUserLocation = false
-    var userLocation:MyLocationMarker? = nil
+    var retrieveLastUserLocation = false
+    var isAdvancedPicker = false
+    var userLocation: MyLocationMarker? = nil
     var dictClusterAnnotation = [String: [StaticGeoPMarker]]()
     var dictIconClusterAnnotation = [String: StaticMarkerData]()
-    var roadMarkerPolyline:TGMarker? = nil
+    var roadMarkerPolyline: TGMarker? = nil
+    var homeMarker: TGMarker? = nil
     var resultFlutter: FlutterResult? = nil
     var methodCall: FlutterMethodCall? = nil
     var uiSingleTapEventMap: UILongPressGestureRecognizer = UILongPressGestureRecognizer()
-    lazy var roadManager:RoadManager = RoadManager()
+    lazy var roadManager: RoadManager = RoadManager()
+
+    let mainView: UIStackView
+    var pickerMarker: UIImageView? = nil
+    var cacheMarkers: [TGMarker] = [TGMarker]()
 
     // var tileRenderer:MKTileOverlayRenderer!
 
@@ -45,6 +52,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
 
         mapView = TGMapView()
         mapView.frame = frame
+        mainView = UIStackView(arrangedSubviews: [mapView])
         //mapview.mapType = MKMapType.standard
         //mapview.isZoomEnabled = true
         //mapview.isScrollEnabled = true
@@ -79,15 +87,18 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             // mapView.requestRender()
 
             let sceneUpdates = [TGSceneUpdate(path: "global.sdk_api_key", value: "qJz9K05vRu6u_tK8H3LmzQ")]
-
-
             let sceneUrl = URL(string: "https://www.nextzen.org/carto/bubble-wrap-style/9/bubble-wrap-style.zip")!
-
+           //var sceneUrl = URL(string: "https://download1490.mediafire.com/79k6g435fjgg/7wmbaskfom9undp/osm-style.zip")
             mapView.loadSceneAsync(from: sceneUrl, with: sceneUpdates)
+
+            //channel.invokeMethod("map#init", arguments: true)
             result(200)
             break
-        case "initPosition":
+        case "initMap":
             initPosition(args: call.arguments, result: result)
+            break;
+        case "changePosition":
+            changePosition(args: call.arguments, result: result)
             break;
         case "currentLocation":
             currentUserLocation()
@@ -98,7 +109,12 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             result(200)
             break;
         case "user#position":
-            /// TODO implement init map with user location without track position
+            retrieveLastUserLocation = true
+            resultFlutter = result
+            locationManager.requestLocation()
+            break;
+        case "goto#position":
+            goToSpecificLocation(call: call, result: result)
             break;
         case "user#pickPosition":
             //let frameV = UIView()
@@ -139,27 +155,44 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             result(200)
             break;
         case "road":
-             drawRoad(call: call) { [unowned self] roadInfo,road,roadData,  error in
-                 if (error != nil) {
-                     result(FlutterError(code: "400", message: "error to draw road", details: nil))
-                 }else {
-                     var newRoad = road
-                     newRoad?.roadData = roadData!
-                     roadManager.drawRoadOnMap(on: newRoad!, for: mapView)
-                     result(roadInfo!.toMap())
-                 }
+            drawRoad(call: call) { [unowned self] roadInfo, road, roadData, error in
+                if (error != nil) {
+                    result(FlutterError(code: "400", message: "error to draw road", details: nil))
+                } else {
+                    var newRoad = road
+                    newRoad?.roadData = roadData!
+                    roadManager.drawRoadOnMap(on: newRoad!, for: mapView)
+                    result(roadInfo!.toMap())
+                }
 
-             }
+            }
             //result(["distance": 0, "duration": 0])
             break;
         case "drawRoad#manually":
-                drawRoadManually(call:call,result:result)
-                break;
+            drawRoadManually(call: call, result: result)
+            break;
+        case "advancedPicker#marker#icon":
+            setCustomIconMarker(call: call, result: result)
+            break;
+        case "advanced#selection":
+            startAdvancedPicker(call: call, result: result)
+            break;
+        case "get#position#advanced#selection":
+            getCenterSelection(call: call, result: result)
+            break;
+        case "confirm#advanced#selection":
+            getCenterSelection(call: call, result: result, isFinished: true)
+            break;
+        case "cancel#advanced#selection":
+            cancelAdvancedPickerMarker()
+            result(200)
+            break;
         default:
             result(nil)
             break;
         }
     }
+
 
 
 
@@ -171,30 +204,57 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
                       MKMapViewDefaultAnnotationViewReuseIdentifier)*/
 
         }
+        //let view = UIStackView(arrangedSubviews: [mapView])
 
-        return mapView
+        return mainView
     }
 
-    private func initPosition(args: Any?, result: FlutterResult) {
+    private func initPosition(args: Any?, result: @escaping FlutterResult) {
         let pointInit = args as! Dictionary<String, Double>
         print(pointInit)
         let location = CLLocationCoordinate2D(latitude: pointInit["lat"]!, longitude: pointInit["lon"]!)
-        mapView.fly(to: TGCameraPosition(center: location, zoom: CGFloat(zoomDefault), bearing: 0, pitch: 0)) { finish in
+
+        // mapView.cameraPosition = TGCameraPosition(center: location, zoom: CGFloat(11), bearing: 0, pitch: 0)
+        channel.invokeMethod("map#init", arguments: true)
+        mapView.fly(to: TGCameraPosition(center: location, zoom: 10.0, bearing: 0, pitch: 0),
+                withDuration: 0.2)
+        /*{ finish in
+            self.
+            // let marker = self.mapView.markerAdd()
+            //self.mapView.markerRemove(marker)
+            result(200)
+        }*/
+        result(200)
+    }
+
+    private func changePosition(args: Any?, result: @escaping FlutterResult) {
+        let pointInit = args as! Dictionary<String, Double>
+        if (homeMarker != nil) {
+            mapView.markerRemove(homeMarker!)
+            mapView.requestRender()
+            homeMarker = nil
+        }
+        let location = CLLocationCoordinate2D(latitude: pointInit["lat"]!, longitude: pointInit["lon"]!)
+        mapView.fly(to: TGCameraPosition(center: location, zoom: mapView.zoom, bearing: 0, pitch: 0), withDuration:0.2) { finish in
             let marker = self.mapView.markerAdd()
             marker.icon = self.markerIcon!
             marker.point = location
             marker.visible = true
             marker.stylingString = "{ style: points, interactive: false,color: white, order: 5000, collide: false }"
+            self.homeMarker = marker
+            result(200)
         }
 
-        //   let geoPoint = GeoPointMap(locationName: "init location", icon: markerIcon, discipline: nil, coordinate: location)
-        //        let span = MKCoordinateSpan(latitudeDelta: 2, longitudeDelta: 2)
-        //    let region = MKCoordinateRegion(center: location, span: span)
-
-        //self.mapView.centerToLocation(geoPoint.location)
-        result(200)
+        //result(200)
     }
+    private func goToSpecificLocation(call: FlutterMethodCall, result: FlutterResult) {
+        let point = call.arguments as! GeoPoint
+        mapView.fly(to: TGCameraPosition(center: point.toLocationCoordinate(), zoom: mapView.zoom, bearing: 0, pitch: 0),
+                 withDuration:0.2)
 
+        result(200)
+
+    }
 
     private func currentUserLocation() {
         locationManager.requestLocation()
@@ -214,34 +274,15 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
     private func zoomMap(_ level: Double) {
         if (level > 0) {
             let cameraPos = TGCameraPosition(center: mapView.position, zoom: mapView.zoom + CGFloat(abs(level)), bearing: 0, pitch: 0)!
-            mapView.fly(to: cameraPos, withSpeed: 25.0)
+            mapView.fly(to: cameraPos, withDuration:0.2)
         } else {
             let cameraPos = TGCameraPosition(center: mapView.position, zoom: mapView.zoom - CGFloat(abs(level)), bearing: 0, pitch: 0)!
-            mapView.fly(to: cameraPos, withSpeed: 25.0)
+            mapView.fly(to: cameraPos, withDuration:0.2)
 
         }
-        /*   var region = MKCoordinateRegion(center: mapView.region.center, span: mapView.region.span)
-
-           if (level > 0) {
-               region.span.latitudeDelta /= level
-               region.span.longitudeDelta /= level
-           } else if (level < 0) {
-               let pLevel = abs(level)
-               region.span.latitudeDelta = min(region.span.latitudeDelta * pLevel, 100.0)
-               region.span.longitudeDelta = min(region.span.longitudeDelta * pLevel, 100.0)
-           }
-           mapView.setRegion(region, animated: true)*/
     }
 
-    private func setupTileRenderer() {
-
-        let template = "https://a.tile.openstreetmap.fr/osm/fr/{z}/{x}/{y}.png"
-
-        let overlay = MKTileOverlay(urlTemplate: template)
-        overlay.canReplaceMapContent = true
-        //self.mapView.addOverlay(overlay, level: .aboveLabels)
-        // self.tileRenderer = MKTileOverlayRenderer(tileOverlay: overlay)
-    }
+    private func setupTileRenderer() {}
 
     private func deactivateTrackMe() {
         canTrackUserLocation = false
@@ -253,14 +294,20 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
         //mapView.showsUserLocation = false
     }
 
+    private func setCustomIconMarker(call: FlutterMethodCall, result: FlutterResult) {
+        let image = convertImage(codeImage: call.arguments as! String)
+        pickerMarker = UIImageView(image: image)
+        result(200)
+    }
+
+
     private func setMarkerStaticGeoPIcon(call: FlutterMethodCall) {
         let args = call.arguments as! [String: Any]
         let id = args["id"] as! String
-        let rgbList: [Int] = (args["color"] as! [Int])
         let icon = convertImage(codeImage: args["bitmap"] as! String)
-        let iconColor = rgbList.toUIColor()//UIColor.init(absoluteRed: rgbList.first!, green: rgbList.last!, blue: rgbList[1])
-        dictIconClusterAnnotation[id] = StaticMarkerData(image: icon!, color: iconColor)
+        dictIconClusterAnnotation[id] = StaticMarkerData(image: icon!)
     }
+
 
     private func setStaticGeoPoint(call: FlutterMethodCall) {
         let args = call.arguments as! [String: Any]
@@ -268,17 +315,13 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
 
 
         let listGeos: [StaticGeoPMarker] = (args["point"] as! [GeoPoint]).map { point -> StaticGeoPMarker in
-            let geo = StaticGeoPMarker(icon: dictIconClusterAnnotation[id]!.image, color: dictIconClusterAnnotation[id]!.color, coordinate: point.toLocationCoordinate())
+            let geo = StaticGeoPMarker(icon: dictIconClusterAnnotation[id]!.image,coordinate: point.toLocationCoordinate())
 
             return geo.addStaticGeosToMapView(for: geo, on: mapView)
         }
 
-        if dictClusterAnnotation.keys.contains(id) {
-            // mapView.removeAnnotations(dictClusterAnnotation[id]!)
-            dictClusterAnnotation[id] = listGeos
-        } else {
-            dictClusterAnnotation[id] = listGeos
-        }
+        dictClusterAnnotation[id] = listGeos
+
 
         // let clusterAnnotation = ClusterMarkerAnnotation(id: id,geos: listGeos)
         //mapView.addAnnotations(listGeos)
@@ -286,52 +329,142 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
 
     }
 
-    private func drawRoad(call: FlutterMethodCall, completion: @escaping (_ roadInfo: RoadInformation?,_ road:Road?,_ roadData:RoadData?, _ error: Error?) -> ()) {
+    public func startAdvancedPicker(call: FlutterMethodCall, result: FlutterResult) {
+        if (!isAdvancedPicker) {
+            isAdvancedPicker = true
+            cacheMarkers += mapView.markers
+            mapView.markerRemoveAll()
+            if (canTrackUserLocation) {
+                deactivateTrackMe()
+            }
+
+            if (pickerMarker == nil) {
+                var image = UIImage(systemName: "markLocation")
+                image = image?.withTintColor(.red)
+                pickerMarker = UIImageView(image: image)
+            }
+            //pickerMarker?.frame = CGRect(x: frame.width/2,y: frame.height/2,width: 32,height: 32)
+            pickerMarker?.center = mainView.center
+            mainView.addSubview(pickerMarker!)
+            result(200)
+        }
+    }
+
+    private func getCenterSelection(call: FlutterMethodCall, result: FlutterResult, isFinished: Bool = false) {
+        if (isAdvancedPicker) {
+            let coordinate = mapView.coordinate(fromViewPosition: mapView.center)
+            result(["lat": coordinate.latitude, "lon": coordinate.longitude])
+            if (isFinished) {
+
+                if (homeMarker != nil) {
+                    homeMarker?.visible = false
+                   let index = cacheMarkers.index(of:homeMarker!)
+                    if(index != nil) {
+                        cacheMarkers.remove(at: index!)
+                    }
+                    homeMarker = nil
+                }
+                let marker = mapView.markerAdd()
+                if(markerIcon != nil){
+                    marker.icon = markerIcon!
+                }
+
+                marker.point = coordinate
+                marker.visible = true
+                marker.stylingString = "{ style: points, interactive: false,color: white, order: 5000, collide: false }"
+                homeMarker = marker
+                cancelAdvancedPickerMarker()
+                isAdvancedPicker = false
+            }
+        }
+    }
+
+    private func cancelAdvancedPickerMarker() {
+        if (isAdvancedPicker) {
+            /// remove picker from parent view
+            pickerMarker?.removeFromSuperview()
+            //pickerMarker = nil
+            cacheMarkers.forEach { marker in
+                let m = mapView.markerAdd()
+                m.stylingString = marker.stylingString
+                if(marker.stylingString.contains("points")) {
+                    m.point = marker.point
+
+                }
+                m.icon = marker.icon
+                if( marker.stylingString.contains("lines")){
+                    m.polyline = marker.polyline
+                }
+                m.visible = marker.visible
+            }
+            cacheMarkers = [TGMarker]()
+            isAdvancedPicker = false
+        }
+    }
+
+
+    private func drawRoad(call: FlutterMethodCall, completion: @escaping (_ roadInfo: RoadInformation?, _ road: Road?, _ roadData: RoadData?, _ error: Error?) -> ()) {
         let args = call.arguments as! [String: Any]
         var points = args["wayPoints"] as! [GeoPoint]
+        points.forEach { p in
+            let markers = mapView.markers.filter { m in
+                m.point == p.toLocationCoordinate()
+            }
+            markers.forEach { m in
+                mapView.markerRemove(m)
+            }
+        }
+        if(roadMarkerPolyline != nil ) {
+            mapView.markerRemove(roadMarkerPolyline!)
+            roadMarkerPolyline = nil
+        }
         var intersectPoint = [GeoPoint]()
-        if(args.keys.contains("middlePoint")){
+        if (args.keys.contains("middlePoint")) {
             intersectPoint = args["middlePoint"] as! [GeoPoint]
             points.insert(contentsOf: intersectPoint, at: 1)
         }
-        var roadColor =  "#ff0000"
-        if(args.keys.contains("roadColor")){
-           roadColor = args["roadColor"] as! String
+        var roadColor = "#ff0000"
+        if (args.keys.contains("roadColor")) {
+            roadColor = args["roadColor"] as! String
         }
         var roadWidth = "5px"
-        if(args.keys.contains("roadWidth")){
-            roadWidth = (args["roadWidth"] as! String) + "px"
+        if (args.keys.contains("roadWidth")) {
+            roadWidth = args["roadWidth"] as! String
         }
 
-        let waysPoint = points.map { point  -> String in
-           let wayP = String(format: "%F,%F", point["lon"]!,point["lat"]!)
-           return  wayP
+        let waysPoint = points.map { point -> String in
+            let wayP = String(format: "%F,%F", point["lon"]!, point["lat"]!)
+            return wayP
         }
-        roadManager.getRoad(wayPoints: waysPoint,typeRoad: RoadType.car){ road in
-               var error: Error? = nil
-                if road==nil {
-                   error =  NSError()
-                    completion(nil,nil,nil,error)
+        roadManager.getRoad(wayPoints: waysPoint, typeRoad: RoadType.car) { road in
+            var error: Error? = nil
+            if road == nil {
+                error = NSError()
+                completion(nil, nil, nil, error)
 
-                }
+            }
             let roadInfo = RoadInformation(distance: road!.distance, seconds: road!.duration)
 
-            completion(roadInfo,road,RoadData(roadColor: roadColor, roadWidth: roadWidth),nil)
+            completion(roadInfo, road, RoadData(roadColor: roadColor, roadWidth: roadWidth), nil)
         }
 
     }
 
     private func drawRoadManually(call: FlutterMethodCall, result: FlutterResult) {
         let args = call.arguments as! [String: Any]
-        var roadEncoded = args["road"] as! String
+        let roadEncoded = args["road"] as! String
 
-        var roadColor =  "#ff0000"
-        if(args.keys.contains("roadColor")){
+        var roadColor = "#ff0000"
+        if (args.keys.contains("roadColor")) {
             roadColor = args["roadColor"] as! String
         }
         var roadWidth = "5px"
-        if(args.keys.contains("roadWidth")){
+        if (args.keys.contains("roadWidth")) {
             roadWidth = "\(args["roadWidth"] as! Double)px"
+        }
+        if(roadMarkerPolyline != nil ) {
+            mapView.markerRemove(roadMarkerPolyline!)
+            roadMarkerPolyline = nil
         }
         var road = Road()
         road.mRouteHigh = roadEncoded
@@ -349,11 +482,12 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
             if let location = locations.last?.coordinate {
                 //mapView.setRegion(region, animated: true)
                 if (canTrackUserLocation) {
-                    if(userLocation == nil ){
+                    if (userLocation == nil) {
                         userLocation = mapView.addUserLocation(for: location, on: mapView)
                     }
-                        userLocation?.marker?.point = location
                     userLocation?.marker?.point = location
+                    //userLocation?.marker?.point = location
+
                     //  mapView.showsUserLocation = true
                     let geoMap = ["lon": location.longitude, "lat": location.latitude]
                     channel.invokeMethod("receiveUserLocation", arguments: geoMap)
@@ -364,6 +498,15 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
                 mapView.flyToUserLocation(for: location)
 
             }
+        } else if (retrieveLastUserLocation) {
+            if let location = locations.last?.coordinate {
+                let geoMap = ["lon": location.longitude, "lat": location.latitude]
+                resultFlutter!(geoMap)
+                retrieveLastUserLocation = false
+                resultFlutter = nil
+            } else {
+                resultFlutter!(nil)
+            }
         }
     }
 
@@ -372,42 +515,51 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
     }
 
     public func mapView(_ mapView: TGMapView, regionDidChangeAnimated animated: Bool) {
-        if( dictClusterAnnotation != nil && !dictClusterAnnotation.isEmpty){
-           for  gStaticMarker in  dictClusterAnnotation {
-               for (i,staticMarker) in gStaticMarker.value.enumerated() {
-                   let m = staticMarker
-                   if(mapView.zoom>12) {
-                       m.marker?.visible = true
-                   }else{
-                       m.marker?.visible = false
-                   }
-                   dictClusterAnnotation[gStaticMarker.key]![i] = m
-               }
-           }
+        if (!dictClusterAnnotation.isEmpty) {
+            for gStaticMarker in dictClusterAnnotation {
+                for (i, staticMarker) in gStaticMarker.value.enumerated() {
+                    let m = staticMarker
+                    if (mapView.zoom > 12) {
+                        m.marker?.visible = true
+                    } else {
+                        m.marker?.visible = false
+                    }
+                    dictClusterAnnotation[gStaticMarker.key]![i] = m
+                }
+            }
+        }
+    }
+
+    public func mapView(_ view: TGMapView!, recognizer: UIGestureRecognizer!,
+                        didRecognizeSingleTapGesture location: CGPoint) {
+        if (resultFlutter != nil && methodCall != nil && methodCall?.method == "user#pickPosition") {
+            var iconM = markerIcon
+            let dict: [String: Any] = methodCall?.arguments as! [String: Any]
+            if let icon = dict["icon"] {
+                iconM = convertImage(codeImage: icon as! String)
+            }
+            let coordinate = view.coordinate(fromViewPosition: location)
+            let geoP = GeoPointMap(icon: iconM!, coordinate: coordinate)
+            geoP.setupMarker(for: geoP, on: view)
+            resultFlutter!(geoP.toMap())
+            methodCall = nil
+        }else{
+            let point = mapView.coordinate(fromViewPosition: location)
+            channel.invokeMethod("receiveSinglePress", arguments: ["lat":point.latitude,"lon":point.longitude])
         }
     }
 
     public func mapView(_ view: TGMapView!, recognizer: UIGestureRecognizer!,
                         didRecognizeLongPressGesture location: CGPoint) {
-        if(resultFlutter != nil && methodCall != nil){
-            var iconM = markerIcon
-            let dict:[String:Any] = methodCall?.arguments as! [String:Any]
-            if let icon = dict["icon"] {
-               iconM =  convertImage(codeImage: icon as! String)
-            }
-            let coordinate = view.coordinate(fromViewPosition: location)
-            let geoP = GeoPointMap(icon: iconM!,coordinate: coordinate)
-            geoP.setupMarker(for: geoP, on: view)
-            resultFlutter!(200)
-            methodCall = nil
-        }
+        let point = mapView.coordinate(fromViewPosition: location)
+        channel.invokeMethod("receiveLongPress", arguments: ["lat":point.latitude,"lon":point.longitude])
 
     }
 
     public func mapView(_ view: TGMapView!, recognizer: UIGestureRecognizer!,
                         shouldRecognizeDoubleTapGesture location: CGPoint) -> Bool {
         let locationMap = view.coordinate(fromViewPosition: location)
-        view.fly(to: TGCameraPosition(center: locationMap, zoom: view.zoom + CGFloat(zoomDefault), bearing: view.bearing, pitch: view.pitch))
+        view.fly(to: TGCameraPosition(center: locationMap, zoom: view.zoom + CGFloat(zoomDefault), bearing: view.bearing, pitch: view.pitch),withDuration:0.2)
         return true
     }
 
@@ -416,41 +568,7 @@ public class MyMapView: NSObject, FlutterPlatformView, CLLocationManagerDelegate
         true
     }
 
-    /*  public func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-          if (annotation is MKUserLocation) {
-              return nil
-          }
-          var viewAnnotation: MKAnnotationView?
-          if let annotationM = annotation as? GeoPointMap {
-              viewAnnotation = annotationM.setupMKAnnotationView(for: annotationM, on: mapView)
-          } else if let staticPoint = annotation as? StaticGeoPMarker {
-              viewAnnotation = staticPoint.setupClusterView(for: staticPoint, on: mapView)
-          }
 
-          return viewAnnotation
-      }*/
-
-    /* public func mapView(
-             _ mapView: MKMapView,
-             rendererFor overlay: MKOverlay
-     ) -> MKOverlayRenderer {
-         if let roadOverlay = overlay as? MKPolyline {
-             let renderer = MKPolylineRenderer(overlay: roadOverlay)
-
-             renderer.strokeColor = .red
-
-             renderer.lineWidth = 15.0
-
-             return renderer
-         }
-
-         guard   let tileOverlay = overlay as? MKTileOverlay else {
-             return MKOverlayRenderer(overlay: overlay)
-         }
-         return MKTileOverlayRenderer(tileOverlay: tileOverlay)
-
-
-     }*/
 }
 
 private extension MKMapView {