+++ lib/src/channel/osm_method_channel.dart
@@ -30,6 +30,10 @@ abstract class EventOSM<T> {
   EventOSM(this.mapId, this.value);
 }
 
+class MapInitialization extends EventOSM<bool> {
+  MapInitialization(int mapId, bool isMapReady) : super(mapId, isMapReady);
+}
+
 class TapEvent extends EventOSM<GeoPoint> {
   TapEvent(int mapId, GeoPoint position) : super(mapId, position);
 }
@@ -81,6 +85,11 @@ class MethodChannelOSM extends OSMPlatform {
   }
 
   @override
+  Stream<MapInitialization> onMapIsReady(int idMap) {
+    return _events(idMap).whereType<MapInitialization>();
+  }
+
+  @override
   Stream<SingleTapEvent> onSinglePressMapClickListener(int idMap) {
     return _events(idMap).whereType<SingleTapEvent>();
   }
@@ -103,6 +112,11 @@ class MethodChannelOSM extends OSMPlatform {
   void setGeoPointHandler(int idMap) async {
     _channels[idMap]!.setMethodCallHandler((call) async {
       switch (call.method) {
+        case "map#init":
+          final result = call.arguments as bool;
+          _streamController.add(MapInitialization(idMap, result));
+
+          break;
         case "receiveLongPress":
           final result = call.arguments;
           _streamController.add(LongTapEvent(idMap, GeoPoint.fromMap(result)));
@@ -132,6 +146,18 @@ class MethodChannelOSM extends OSMPlatform {
   }
 
   @override
+  Future<void> initMap(
+    int idOSM,
+    GeoPoint point,
+  ) async {
+    Map requestData = {"lon": point.longitude, "lat": point.latitude};
+    await _channels[idOSM]!.invokeMethod(
+      "initMap",
+      requestData,
+    );
+  }
+
+  @override
   Future<void> currentLocation(int? idOSM) async {
     try {
       await _channels[idOSM]!.invokeMethod("currentLocation", null);
@@ -155,7 +181,7 @@ class MethodChannelOSM extends OSMPlatform {
   Future<void> addPosition(int idOSM, GeoPoint p) async {
     Map requestData = {"lon": p.longitude, "lat": p.latitude};
     await _channels[idOSM]!.invokeMethod(
-      "initPosition",
+      "changePosition",
       requestData,
     );
   }
@@ -176,18 +202,18 @@ class MethodChannelOSM extends OSMPlatform {
   Future<void> customMarkerStaticPosition(
     int idOSM,
     GlobalKey? globalKey,
-    String id, {
-    Color? colorIcon,
-  }) async {
+    String id,
+  ) async {
     Uint8List icon = await _capturePng(globalKey!);
+    String iconIOS = "";
+    if (Platform.isIOS) {
+      iconIOS = icon.convertToString();
+    }
     var args = {
       "id": id,
-      "bitmap": icon,
+      "bitmap": Platform.isIOS ? iconIOS : icon,
     };
-    if (Platform.isIOS && colorIcon != null) {
-      args.addAll(colorIcon.toMap("color"));
-      args["bitmap"] = icon.convertToString();
-    }
+
     await _channels[idOSM]!.invokeMethod(
       "staticPosition#IconMarker",
       args,
@@ -223,7 +249,7 @@ class MethodChannelOSM extends OSMPlatform {
         args.addAll(roadOption.roadColor!.toHexMap("roadColor"));
       }
       if (roadOption.roadWidth != null) {
-        args.addAll({"roadWidth": roadOption.roadWidth});
+        args.addAll({"roadWidth": "${roadOption.roadWidth}px"});
       }
     } else {
       if (roadOption.roadColor != null) {
@@ -261,7 +287,7 @@ class MethodChannelOSM extends OSMPlatform {
     Map args = {};
     if (key != null) {
       bitmap = await _capturePng(key);
-      args.addAll({"icon": bitmap});
+      args.addAll({"icon": Platform.isIOS ? bitmap.convertToString() : bitmap});
     }
     if (imageURL.isNotEmpty) {
       args.addAll({"imageURL": imageURL});
@@ -469,7 +495,11 @@ class MethodChannelOSM extends OSMPlatform {
       "road": encodedCoordinates,
       "roadWidth": width,
     };
-    data.addAll(roadColor.toMap("roadColor"));
+    if (Platform.isIOS) {
+      data.addAll(roadColor.toHexMap("roadColor"));
+    } else {
+      data.addAll(roadColor.toMap("roadColor"));
+    }
 
     await _channels[idOSM]!.invokeMethod(
       "drawRoad#manually",