+++ lib/src/interface_osm/osm_interface.dart
@@ -59,9 +59,8 @@ abstract class OSMPlatform extends BaseOsmPlatform {
   Future<void> customMarkerStaticPosition(
     int idOSM,
     GlobalKey? globalKey,
-    String id, {
-    Color? colorIcon,
-  });
+    String id,
+  );
 
   Future<void> staticPosition(
     int idOSM,